requirements.txt
Django==2.2.17
django-cors-headers==3.6.0
djangorestframework==3.12.2
gunicorn==20.0.4
hedju==0.1.0
psycopg2==2.8.6
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.2.0
